requirements.txt
aiogram==3.2.0
python-dotenv
aiofiles
httpx

